+++ src/services/deliveryDistanceService.ts
@@ -1,18 +1,9 @@
 import axios from 'axios'
 import config from '../config'
-
 const googleMapsApiKey = config.GOOGLE_MAPS_API_KEY
-
-export interface Geometry {
-  location: {
-    lat: number
-    lng: number
-  }
-}
 
 export interface LocationDetails {
   geometry: Geometry
-  // Include other properties as needed
 }
 export const calculateDistanceService = async (
   pickupLocation: LocationDetails,
@@ -65,47 +56,5 @@ export const calculateDistanceService = async (
       console.error('Unknown error calculating distance:', error)
       throw new Error('Unknown error occurred while calculating distance')
     }
-  }
-}
-
-interface DeliveryRequestBody {
-  pickupLocation: LocationDetails
-  dropoffLocation: LocationDetails
-  package_size: 'small' | 'medium' | 'large'
-  delivery_type: 'standard' | 'express'
-}
-
-export const calculateDeliveryCostService = async (
-  deliveryRequest: DeliveryRequestBody
-): Promise<number | null> => {
-  const { pickupLocation, dropoffLocation, package_size, delivery_type } =
-    deliveryRequest
-
-  //TODO: Replace this with the formular for calculating distance from the demo
-  const distance = await calculateDistanceService(
-    pickupLocation,
-    dropoffLocation
-  )
-  if (distance) {
-    let cost = 10 // Base cost
-    cost += distance * 0.5 // Add cost based on distance
-
-    switch (package_size) {
-      case 'medium':
-        cost += 5
-        break
-      case 'large':
-        cost += 10
-        break
-      // No additional cost for 'small'
-    }
-
-    if (delivery_type === 'express') {
-      cost *= 1.5 // Express delivery costs 50% more
-    }
-
-    return cost
-  } else {
-    return null
   }
 }
